+++ src/compiler/typescript_to_rust/azle_vm_value_derive/src/derive_try_from_vm_value/mod.rs
@@ -1,0 +1,3 @@
+pub mod derive_try_from_vm_value_enum;
+pub mod derive_try_from_vm_value_struct;
+pub mod derive_try_from_vm_value_vec;
