+++ src/compiler/typescript_to_rust/azle_vm_value_derive/src/derive_try_into_vm_value/derive_try_into_vm_value_enum.rs
@@ -2,7 +2,7 @@ use proc_macro2::Ident;
 use quote::{format_ident, quote};
 use syn::{DataEnum, Field, Fields, Generics};
 
-pub fn derive_try_into_vm_value_enum(
+pub fn generate(
     enum_name: &Ident,
     data_enum: &DataEnum,
     generics: &Generics,
