+++ src/compiler/typescript_to_rust/azle_vm_value_derive/src/derive_try_into_vm_value/mod.rs
@@ -1,0 +1,2 @@
+pub mod derive_try_into_vm_value_enum;
+pub mod derive_try_into_vm_value_struct;
