+++ src/compiler/typescript_to_rust/azle_vm_value_derive/src/traits/mod.rs
@@ -1,0 +1,5 @@
+pub use map_to::MapTo;
+pub use try_get_ident::TryGetIdent;
+
+pub mod map_to;
+pub mod try_get_ident;
