+++ src/compiler/typescript_to_rust/azle_vm_value_derive/src/traits/try_get_ident.rs
@@ -1,0 +1,30 @@
+use proc_macro2::{Ident, Span};
+use syn::{Error, Field};
+
+pub trait TryGetIdent {
+    fn try_get_ident(
+        &self,
+        enum_name: &Ident,
+        variant_name: &Ident,
+        field_index: usize,
+    ) -> Result<&Ident, Error>;
+}
+
+impl TryGetIdent for Field {
+    fn try_get_ident(
+        &self,
+        enum_name: &Ident,
+        variant_name: &Ident,
+        field_index: usize,
+    ) -> Result<&Ident, Error> {
+        self.ident.as_ref().ok_or_else(|| {
+            Error::new(
+                Span::call_site(),
+                format!(
+                    "Internal Error: expected field {field_index} in \
+                        {enum_name}::{variant_name} to have a name but received None"
+                ),
+            )
+        })
+    }
+}
